src/kafka_utils/kafka_admin.py
"""Kafka administration for topic creation and management."""

from kafka.admin import KafkaAdminClient, NewTopic

class KafkaTopicManager:
    """Manage Kafka topics for the trading system."""
    
    TOPICS_CONFIG = {
        "market_data": {"partitions": 12, "replication_factor": 3},
        "indicators": {"partitions": 6, "replication_factor": 3},
        "signals": {"partitions": 3, "replication_factor": 3},
        "orders": {"partitions": 3, "replication_factor": 3},
        "risk_events": {"partitions": 3, "replication_factor": 3},
        "logs": {"partitions": 6, "replication_factor": 3},
    }
    
    def __init__(self, bootstrap_servers: str):
        self.admin_client = KafkaAdminClient(
            bootstrap_servers=bootstrap_servers,
            client_id="trading_system_admin"
        )
    
    def create_topics(self):
        """Create all required topics."""
        pass
