requirements.txt
# Finstock Trading System — Root Dependencies
# For local development and testing. Each service has its own requirements.txt.

# Data Sources
vnstock>=3.2.6

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# Technical Analysis
ta>=0.11.0

# Messaging
kafka-python>=2.0.2
python-snappy>=0.6.1

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Data Validation
pydantic>=2.5.0

# Object Storage
minio>=7.2.0
pyarrow>=14.0.0

# Machine Learning (optimization)
scikit-learn>=1.3.0

# HTTP
requests>=2.28.0

# Configuration & Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.0
pytz>=2024.1
structlog>=23.1.0

# Development & Testing
pytest>=7.0.0
pytest-cov>=4.0.0
